Update and rename lotes de Enteros_promedioMaximo.c to batchesOfNumbers.c
The user is asked to enter N numbers equal or greater to zero. The average of those numbers is calculated. Every time a zero is entered, the program separates the previous numbers creating a batch, and then the process is repeated to make more groups of numbers, each with it´s average, until the N numbers are input. Finally, the program prints the number of the group with it´s highest average.
Please notice that the group (or batch) is only created after a zero is input.